Looking at the output next to the beta values, it seems like the problem is not one of optimization, but of functional estimation.
I suspect that only certain training vectors are intended to be used to generate the estimate, and that these are selected based on
their beta value.  When gamma is extrememly low, beta values are uniform and the estimate matches perfectly.  As gamma increases,
beta becomes more varied, and as beta diverges from it's minimum, the estimate diverges from the actual.

It's possible that beta is describing the 'support vectors', and that these spike indicate representational features of the dataset.